+++ packages/web-sdk/src/instrumentations/errors/index.ts
@@ -1,5 +1,3 @@
 export { ErrorsInstrumentation } from './instrumentation';
 
-export { buildStackFrame, getDataFromSafariExtensions, getStackFramesFromError, parseStacktrace } from './stackFrames';
-
 export type { ErrorEvent, ExtendedPromiseRejectionEvent } from './types';
