--- packages/web-sdk/src/instrumentations/errors/stackFrames/parseStacktrace.ts
@@ -1,9 +1,0 @@
-import type { ExtendedError, Stacktrace } from '@grafana/faro-core';
-
-import { getStackFramesFromError } from './getStackFramesFromError';
-
-export function parseStacktrace(error: ExtendedError): Stacktrace {
-  return {
-    frames: getStackFramesFromError(error),
-  };
-}
