+++ packages/web-sdk/src/instrumentations/errors/types.ts
@@ -1,3 +1,5 @@
+import type { StackframeParserOptions } from "../../utils/stackFrames/types";
+
 export interface ExtendedPromiseRejectionEvent extends PromiseRejectionEvent {
   detail?: {
     reason: PromiseRejectionEvent['reason'];
@@ -7,3 +9,7 @@ export interface ExtendedPromiseRejectionEvent extends PromiseRejectionEvent {
 export type ErrorEvent = (Error | Event) & {
   error?: Error;
 };
+
+export type ErrorInstrumentationOptions = {
+  stackframeParserOptions?: StackframeParserOptions
+}
