+++ packages/web-sdk/src/instrumentations/index.ts
@@ -4,12 +4,9 @@ export { ConsoleInstrumentation } from './console';
 export type { ConsoleInstrumentationOptions } from './console';
 
 export {
-  buildStackFrame,
   ErrorsInstrumentation,
-  getDataFromSafariExtensions,
-  getStackFramesFromError,
-  parseStacktrace,
 } from './errors';
+
 export type { ErrorEvent, ExtendedPromiseRejectionEvent } from './errors';
 
 export { ViewInstrumentation } from './view';
