+++ packages/web-sdk/src/utils/stackFrames/buildStackFrame.ts
