+++ packages/web-sdk/src/utils/stackFrames/const.ts
