+++ packages/web-sdk/src/utils/stackFrames/getDataFromSafariExtensions.ts
