+++ packages/web-sdk/src/utils/stackFrames/getStackFramesFromError.test.ts
