+++ packages/web-sdk/src/utils/stackFrames/getStackFramesFromError.ts
@@ -1,5 +1,6 @@
 import { isNumber } from '@grafana/faro-core';
 import type { ExceptionStackFrame, ExtendedError } from '@grafana/faro-core';
+import type { StackframeParserOptions } from './types';
 
 import { buildStackFrame } from './buildStackFrame';
 import {
@@ -17,7 +18,7 @@ import {
 } from './const';
 import { getDataFromSafariExtensions } from './getDataFromSafariExtensions';
 
-export function getStackFramesFromError(error: ExtendedError): ExceptionStackFrame[] {
+export function getStackFramesFromError(error: ExtendedError, options: StackframeParserOptions = {}): ExceptionStackFrame[] {
   let lines: string[] = [];
 
   if (error.stacktrace) {
@@ -32,6 +33,13 @@ export function getStackFramesFromError(error: ExtendedError): ExceptionStackFra
     let filename: string | undefined;
     let lineno: string | undefined;
     let colno: string | undefined;
+
+    // skip attempting to parse stack frames over the limit, if it is set and above 0
+    if(options.maximumLineLength !== undefined && options.maximumLineLength > 0) {
+      if(line.length > options.maximumLineLength) {
+        return acc;
+      }
+    }
 
     if ((parts = webkitLineRegex.exec(line))) {
       func = parts[1];
