+++ packages/web-sdk/src/utils/stackFrames/parseStacktrace.ts
@@ -1,0 +1,20 @@
+import type { ExtendedError, Stacktrace, StacktraceParser } from '@grafana/faro-core';
+
+import type { StackframeParserOptions } from './types';
+
+import { getStackFramesFromError } from './getStackFramesFromError';
+
+
+export function parseStacktrace(error: ExtendedError): Stacktrace {
+  return {
+    frames: getStackFramesFromError(error),
+  };
+}
+
+export function newStackTraceParser(options?: StackframeParserOptions): StacktraceParser {
+  return (error: ExtendedError)=>{
+    return {
+      frames: getStackFramesFromError(error, options),
+    }
+  }
+}
