+++ packages/web-sdk/src/utils/stackFrames/types.ts
@@ -1,0 +1,3 @@
+export interface StackframeParserOptions {
+  maximumLineLength?: number;
+}
